agentcircle-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
supabase==2.0.3
python-dotenv==1.0.0
apscheduler==3.10.4
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.1.0
httpx==0.25.2
